Added vgg16 model
I used Transfer learning with the state-of-art CNN architecture VGG16 model and it gave me accuracy of 99% in training and 94% in validation. Early i got 84% training and 81% validation accuracy without using the transfer learning. It shows importance of transfer learning in Computer Vision. Due to the less number of data overfitting is occurs in small scale and we can get rid of it by using dropout and L2 regularization and also by adding the data.